FastAPI_Backend/requirements.txt
# --- Core Framework ---
fastapi
uvicorn

# --- LangChain + Gemini (keep your versions) ---
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-google-genai==0.0.6

# --- Environment ---
python-dotenv

# --- PDF Generation (you had this) ---
fpdf==1.7.2

# --- Authentication & Security ---
passlib[bcrypt]==1.7.4
PyJWT==2.9.0

# --- Database (MongoDB) ---
motor==3.5.1
pymongo==4.10.1

# --- Data Validation ---
pydantic==2.9.2
email-validator==2.2.0

# --- Networking / Requests ---
requests==2.32.3
httpx==0.27.2
anyio==4.4.0
